lambda/requirements.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0
aiobotocore>=2.7.0
boto3>=1.34.0
pyyaml>=6.0.0
